[init] 초기화 : 기존 sprint1 유지하고 모범답안으로 sprint2  진행
