--- src/main/java/com/sprint/mission/model/entity/BaseTimeEntity.java
@@ -1,28 +1,0 @@
-package com.sprint.mission.model.entity;
-
-
-
-public abstract class BaseTimeEntity {
-
-    private long created;
-    private long updatedAt;
-
-    public BaseTimeEntity() {
-        this.created = System.currentTimeMillis();
-        this.updatedAt = this.created;
-    }
-
-    public long getUpdatedAt() {
-        return updatedAt;
-    }
-
-
-    public long getCreated() {
-        return created;
-    }
-
-    public void updateTimestamp() {
-        this.updatedAt = System.currentTimeMillis();
-    }
-
-}
