--- src/main/java/com/sprint/mission/model/entity/Channel.java
@@ -1,12 +1,0 @@
-package com.sprint.mission.model.entity;
-
-public class Channel {
-
-    /**
-     *  NOTE : MYSQl 기준
-     */
-    public String channelType;
-
-
-
-}
