--- src/main/java/com/sprint/mission/view/input/UserInput.java
@@ -1,61 +1,0 @@
-package com.sprint.mission.view.input;
-
-import com.sprint.mission.service.UserService;
-import java.util.Scanner;
-
-public class UserInput {
-
-
-    private final Scanner scanner;
-
-    private final UserService userService;
-
-
-    public UserInput(UserService userService) {
-        this.userService = userService;
-        this.scanner = new Scanner(System.in);
-    }
-
-    public void creatInput() {
-        System.out.println("사용자을 입력해주세요");
-        String username = scanner.nextLine().trim();
-
-        System.out.print("이메일을 입력하세요: ");
-        String email = scanner.nextLine().trim();
-
-        System.out.print("비밀번호를 입력하세요: ");
-        String password = scanner.nextLine().trim();
-
-        userService.createUser(username, email, password);
-    }
-
-    public void updateInput() {
-
-        System.out.print("사용 중이신 이메일을 입력해주세요(이메일은 변경이 불가능 합니다): ");
-        String email = scanner.nextLine().trim();
-
-        System.out.print("변경 할 이름을 입력해주세요: ");
-        String username = scanner.nextLine().trim();
-
-        System.out.print("변경 할 비밀번호 입력해주세요: ");
-        String password = scanner.nextLine().trim();
-        userService.updateUser(email, username, password);
-    }
-
-    /**
-     *  TODO ID 으로  입력 받기 나중에 구현
-     */
-    public void getAllInput() {
-        userService.getAllUser();
-    }
-
-
-    public void getEmailInput() {
-        System.out.print("찾을 이메일을 입력하세요: ");
-        String email = scanner.nextLine().trim();
-
-        userService.getEmailUser(email);
-    }
-
-
-}
