--- src/main/java/com/sprint/mission/view/output/MessageOutput.java
@@ -1,14 +1,0 @@
-package com.sprint.mission.view.output;
-
-import com.sprint.mission.model.entity.User;
-
-public class MessageOutput {
-
-    public void checkOutput(User user) {
-        System.out.println(user.getUsername() + "님, 채팅을 시작합니다. ('종료' 입력 시 종료)");
-    }
-
-    public void displayMessage(String username, String message) {
-        System.out.println(username + ": " + message);
-    }
-}
