+++ src/main/java/com/sprint/sprint1/mission/model/entity/Channel.java
@@ -1,0 +1,12 @@
+package com.sprint.sprint1.mission.model.entity;
+
+public class Channel {
+
+    /**
+     *  NOTE : MYSQl 기준
+     */
+    public String channelType;
+
+
+
+}
