+++ src/main/java/com/sprint/sprint1/mission/model/entity/Message.java
@@ -1,0 +1,28 @@
+package com.sprint.sprint1.mission.model.entity;
+
+public class Message extends BaseTimeEntity{
+
+    private final String channel;
+    private final User sender;
+    private String content;
+
+
+    public Message(String channel, User sender, String content) {
+        super();
+        this.channel = channel;
+        this.sender = sender;
+        this.content = content;
+    }
+
+    public String getChannel() {
+        return channel;
+    }
+
+    public User getSender() {
+        return sender;
+    }
+
+    public String getContent() {
+        return content;
+    }
+}
