+++ src/main/java/com/sprint/sprint1/mission/service/UserService.java
@@ -1,0 +1,13 @@
+package com.sprint.sprint1.mission.service;
+
+public interface UserService {
+
+
+    void createUser(String username, String email, String password);
+
+    void updateUser(String email, String username, String password);
+
+    void getAllUser();
+
+    void getEmailUser(String email);
+}
