+++ src/main/java/com/sprint/sprint2/discodeit/entity/ChannelType.java
@@ -1,0 +1,6 @@
+package com.sprint.sprint2.discodeit.entity;
+
+public enum ChannelType {
+    PUBLIC,
+    PRIVATE,
+}
