+++ src/main/java/com/sprint/sprint2/discodeit/service/MessageService.java
@@ -1,0 +1,14 @@
+package com.sprint.sprint2.discodeit.service;
+
+import com.sprint.sprint2.discodeit.entity.Message;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface MessageService {
+    Message create(String content, UUID channelId, UUID authorId);
+    Message find(UUID messageId);
+    List<Message> findAll();
+    Message update(UUID messageId, String newContent);
+    void delete(UUID messageId);
+}
