+++ src/main/java/com/sprint/sprint2/discodeit/service/UserService.java
@@ -1,0 +1,14 @@
+package com.sprint.sprint2.discodeit.service;
+
+import com.sprint.sprint2.discodeit.entity.User;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface UserService {
+    User create(String username, String email, String password);
+    User find(UUID userId);
+    List<User> findAll();
+    User update(UUID userId, String newUsername, String newEmail, String newPassword);
+    void delete(UUID userId);
+}
