+++ src/main/java/com/sprint/sprint2/discodeit/service/jcf/JCFChannelService.java
@@ -1,0 +1,52 @@
+package com.sprint.sprint2.discodeit.service.jcf;
+
+import com.sprint.sprint2.discodeit.entity.Channel;
+import com.sprint.sprint2.discodeit.entity.ChannelType;
+import com.sprint.sprint2.discodeit.service.ChannelService;
+import java.util.*;
+
+public class JCFChannelService implements ChannelService {
+    private final Map<UUID, Channel> data;
+
+    public JCFChannelService() {
+        this.data = new HashMap<>();
+    }
+
+    @Override
+    public Channel create(ChannelType type, String name, String description) {
+        Channel channel = new Channel(type, name, description);
+        this.data.put(channel.getId(), channel);
+
+        return channel;
+    }
+
+    @Override
+    public Channel find(UUID channelId) {
+        Channel channelNullable = this.data.get(channelId);
+        return Optional.ofNullable(channelNullable)
+                        .orElseThrow(() -> new NoSuchElementException("Channel with id " + channelId + " not found"));
+    }
+
+    @Override
+    public List<Channel> findAll() {
+        return this.data.values().stream().toList();
+    }
+
+    @Override
+    public Channel update(UUID channelId, String newName, String newDescription) {
+        Channel channelNullable = this.data.get(channelId);
+        Channel channel = Optional.ofNullable(channelNullable)
+                .orElseThrow(() -> new NoSuchElementException("Channel with id " + channelId + " not found"));
+        channel.update(newName, newDescription);
+
+        return channel;
+    }
+
+    @Override
+    public void delete(UUID channelId) {
+        if (!this.data.containsKey(channelId)) {
+            throw new NoSuchElementException("Channel with id " + channelId + " not found");
+        }
+        this.data.remove(channelId);
+    }
+}
